+++ libs/admin/src/lib/admin-panel/components/guest-table/guest-table.component.html
@@ -39,7 +39,7 @@
   </ng-template>
   <ng-template colRef="movie" let-movie>
     <div *ngIf="movie else noMovie" fxLayout="row">
-      <ng-container *ngIf="movie.main.poster && movie.main.poste.media as poster;">
+      <ng-container *ngIf="movie.main.poster && movie.main.poster.media as poster;">
         <bf-img class="poster" [ref]="poster" asset="empty_poster.webp"></bf-img>
       </ng-container>
       <a [routerLink]="['/c/o/admin/panel/movie', movie.id]">
