+++ libs/movie/src/lib/movie/+state/movie.service.ts
@@ -15,7 +15,6 @@ import { RouterQuery } from '@datorama/akita-ng-router-store';
 import { OrganizationService } from '@blockframes/organization/+state/organization.service';
 import { UserService } from '@blockframes/user/+state/user.service';
 import { MediaService } from '@blockframes/media/+state/media.service';
-// import { extractToBeUpdatedMedia } from '@blockframes/media/+state/media.model';
 import { firestore } from 'firebase/app';
 import { createMovieAppAccess, getCurrentApp } from '@blockframes/utils/apps';
 
@@ -52,12 +51,6 @@ export class MovieService extends CollectionService<MovieState> {
       movie.id = await this.add(cleanModel(movie), { write: tx });
     });
     return movie;
-  }
-
-  save(movie: Movie) {
-    return this.runTransaction(async write => {
-      return this.update(movie, { write });
-    });
   }
 
   async onCreate(movie: Movie, { write }: WriteOptions) {
