+++ libs/user/src/lib/auth/pages/profile-view/profile-view.component.html
@@ -16,7 +16,7 @@
   </header>
 </ng-container>
 <section>
-  <auth-form-profile [profileForm]="profileForm"></auth-form-profile>
+  <auth-form-profile [form]="profileForm"></auth-form-profile>
   <auth-form-password [form]="passwordForm"></auth-form-password>
 </section>
 <footer fxLayout="row" fxLayoutAlign="center center">
